Move classes '.core' to '.server.jetty'
