+++ jdisc_http_service/src/main/java/com/yahoo/jdisc/http/server/jetty/HttpServletRequestUtils.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.jdisc.http.core;
+package com.yahoo.jdisc.http.server.jetty;
 
 import org.eclipse.jetty.server.HttpConnection;
 import org.eclipse.jetty.server.ServerConnector;
